+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Release Notes for Helpdesk Support
 
+## 1.1.0 - 2019-11-19
+### Added
+- Disabled the "My Tickets" Inbox option for Teamwork (not really an Inbox, just a collection of tickets)
+- Teamwork Desk now allows for adding attachments when creating a new ticket
+
 ## 1.0.0 - 2019-11-18
 ### Initial release